Bump HOSContract version to 0.3 for clock backwards compatibility
Additional testing of opencab#19 revealed that although `Parcel` is capable
of reading a single parceled object from a newer version that adds
additional fields to the end (in this case, `durationSeconds`), since
the `Clock` objects are in an array, the increased length of the
unconsumed field throws off the offset of subsequent objects in the
array and leads to a crash inside the `Parcel` code.

To avoid causing such a crash in a pre-existing HOS consumer that only
understands version 0.2 clocks (no `durationSeconds`) when it encounters
values from a version 0.3 provider, we are bumping the version number so
that:

  1. New consumers MUST explicitly indicate support for `durationClocks` by
     passing `"0.3"` to the `version` parameter of `ContractProvider.call`
     when fetching HOS status.
  2. New providers MUST NOT return a version 0.3 clock (with a
     `durationSeconds` value) if the calling consumer has not passed a
     version equal to or greater than `"0.3"`. New providers MAY return
     an error if they do not need or want to support older consumers.
  3. All new providers and consumers SHOULD support version 0.3 so that
     support for 0.2 can eventually be dropped.

In addition, the fact that `Parcel` can't natively handle new values
in a backwards-compatible way means we'll stop using it for all future
contracts or contract versions, with the exception of the in-flight
opencab#17 work.

We need to define "greater than" for version numbers, presumably using
non-lexical comparison.

Co-authored-by: Victor Ceban <victor.ceban@eleostech.com>