Update examples to show how providers need to enumerate receiver packages
The previous sample consumer had its receiver classes all named at the top
level of the app's package. This resulted in one implementation that
inadvertently depended on the receiver class not being in a sub-package,
which broke compatibility.

The existing sample provider had code to enumerate packages, but this
updates the consumer to actually exercise this by moving the receiver
package to a sub-package. It also unifies this logic across all
broadcast types.